SqrtX69 solution 2 added.

+++ src/main/java/tradition150/SqrtX69.java
@@ -42,6 +42,6 @@ public class SqrtX69 {
             return 0;
         }
         int ans = (int) Math.exp(0.5 * Math.log(x));
-        return (ans + 1) * (ans + 1) <= x ? ans + 1 : ans;
+        return (long) (ans + 1) * (ans + 1) <= x ? ans + 1 : ans;
     }
 }
